+++ src/mantarray_desktop_app/constants.py
@@ -85,7 +85,6 @@ VALID_CONFIG_SETTINGS = frozenset(
         "customer_account_uuid",
         "user_account_id",
         "customer_pass_key",
-        "user_account_id",
         "recording_directory",
         "auto_upload",
         "auto_delete",